Core: tidy

+++ src/core/processors/poolprocessor.cpp
@@ -46,7 +46,7 @@ void pool::Progress::operator()(double progress) const noexcept {
 }
 
 void pool::Progress::operator()(size_t i, size_t max) const noexcept {
-    state_.setProgress(id_, static_cast<double>(i) / max);
+    state_.setProgress(id_, static_cast<double>(i) / static_cast<double>(max));
 }
 
 void pool::detail::State::setProgress(size_t id, double newProgress) {
@@ -55,8 +55,8 @@ void pool::detail::State::setProgress(size_t id, double newProgress) {
     progress[id] = newProgress;
 
     if (!progressUpdate.valid() || util::is_future_ready(progressUpdate)) {
-        const auto total =
-            std::accumulate(progress.begin(), progress.end(), 0.0) / progress.size();
+        const auto total = std::accumulate(progress.begin(), progress.end(), 0.0) /
+                           static_cast<double>(progress.size());
         progressUpdate = dispatchFront([this, poolProcessor = processor, total]() {
             if (auto p = poolProcessor.lock()) {
                 p->progress(this, total);
